# openapi.yaml
openapi: 3.0.3
info:
  title: Expense Tracker API
  version: 1.0.0
  description: API for managing accounts, categories, debtors, transactions, and authentication.
servers:
  - url: /api
paths:
...
  /dtrans:
    get:
      summary: Get all debtor transactions
      responses:
        '200':
          description: List of debtor transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DTrans'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new debtor transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DTrans'
      responses:
        '201':
          description: Debtor transaction created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/DTrans'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /transactions/{id}:
    get:
      summary: Get transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Transaction'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Account:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        balance: { type: number }
    Category:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
...
    DTrans:
      type: object
      properties:
        _id: { type: string }
        debtorId: { type: string }
        amount: { type: number }
        date: { type: string, format: date-time }
        expectedReturnDate: { type: string, format: date-time }
    Transaction:
      type: object
      properties:
        _id: { type: string }
        type: { type: string, enum: [income, expense] }
        amount: { type: number }
        date: { type: string, format: date-time }
        account:
          $ref: '#/components/schemas/Account'
        category:
          $ref: '#/components/schemas/Category'
        status: { type: string, enum: [open, completed, pending] }
        to: { type: string }
    Error:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string